1010: Moving datamodile classes under v3_1_10 pkg
https://github.com/SecureApiGateway/SecureApiGateway/issues/1010
