+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-datamodel/src/main/java/com/forgerock/sapi/gateway/rcs/conent/store/datamodel/payment/domestic/v3_1_10/AuthoriseDomesticPaymentConsentRequest.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.rcs.conent.store.datamodel.payment.domestic;
+package com.forgerock.sapi.gateway.rcs.conent.store.datamodel.payment.domestic.v3_1_10;
 
 import javax.validation.constraints.NotNull;
 
 import org.springframework.validation.annotation.Validated;
 
 @Validated
-public class RejectDomesticPaymentConsentRequest {
+public class AuthoriseDomesticPaymentConsentRequest {
 
     @NotNull
     private String apiClientId;
@@ -31,7 +31,10 @@ public class RejectDomesticPaymentConsentRequest {
     @NotNull
     private String resourceOwnerId;
 
-    public RejectDomesticPaymentConsentRequest(){
+    @NotNull
+    private String authorisedDebtorAccountId;
+
+    public AuthoriseDomesticPaymentConsentRequest() {
     }
 
     public String getConsentId() {
@@ -48,6 +51,14 @@ public class RejectDomesticPaymentConsentRequest {
 
     public void setResourceOwnerId(String resourceOwnerId) {
         this.resourceOwnerId = resourceOwnerId;
+    }
+
+    public String getAuthorisedDebtorAccountId() {
+        return authorisedDebtorAccountId;
+    }
+
+    public void setAuthorisedDebtorAccountId(String authorisedDebtorAccountId) {
+        this.authorisedDebtorAccountId = authorisedDebtorAccountId;
     }
 
     public String getApiClientId() {
